refactor(srv): split service.rs into concern submodules
