+++ .changesets/1782953967-refactor-srv-split-service-into-concern-submodules-bxdr.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(srv): split service into concern submodules
