+++ agentmux-srv/src/server/service/client.rs
@@ -1,0 +1,63 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+//! `client` service handler.
+
+use crate::backend::obj::*;
+use crate::backend::service::{self, WebCallType, WebReturnType};
+use crate::backend::wcore;
+
+use super::super::AppState;
+
+pub(super) async fn handle_client_service(state: &AppState, call: &WebCallType) -> WebReturnType {
+    let store = &state.wstore;
+    let args = &call.args;
+    match call.method.as_str() {
+        "GetClientData" => match wcore::get_client(store) {
+            Ok(client) => {
+                WebReturnType::success(serde_json::to_value(&client).unwrap_or_default())
+            }
+            Err(e) => WebReturnType::error(e.to_string()),
+        },
+        "GetTab" => {
+            let tab_id: String = match service::get_arg(args, 0) {
+                Ok(v) => v,
+                Err(e) => return WebReturnType::error(e),
+            };
+            match store.must_get::<Tab>(&tab_id) {
+                Ok(tab) => WebReturnType::success(serde_json::to_value(&tab).unwrap_or_default()),
+                Err(e) => WebReturnType::error(e.to_string()),
+            }
+        }
+        "FocusWindow" => {
+            let window_id: String = match service::get_arg(args, 0) {
+                Ok(v) => v,
+                Err(e) => return WebReturnType::error(e),
+            };
+            match wcore::focus_window(store, &window_id) {
+                Ok(()) => WebReturnType::success_empty(),
+                Err(e) => WebReturnType::error(e.to_string()),
+            }
+        }
+        "AgreeTos" => match wcore::get_client(store) {
+            Ok(mut client) => {
+                client.tosagreed = chrono::Utc::now().timestamp_millis();
+                match store.update(&mut client) {
+                    Ok(_) => WebReturnType::success_empty(),
+                    Err(e) => WebReturnType::error(e.to_string()),
+                }
+            }
+            Err(e) => WebReturnType::error(e.to_string()),
+        },
+        "GetAllConnStatus" => {
+            // Return empty — connection manager not yet wired
+            // Go returns success with no data (nil slice omitted by omitempty)
+            WebReturnType::success_empty()
+        }
+        "TelemetryUpdate" => {
+            // Accept but ignore — telemetry not implemented
+            WebReturnType::success_empty()
+        }
+        _ => WebReturnType::error(format!("unknown client method: {}", call.method)),
+    }
+}
